requirements.txt
amqp==2.5.2
arrow==0.15.8
asgiref==3.2.7
asn1crypto==0.24.0
audioread==2.1.8
beautifulsoup4==4.8.0
billiard==3.6.3.0
blessed==1.17.9
bs4==0.0.1
cachetools==4.1.0
celery==4.4.2
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
click==7.1.1
cryptography==2.7
cycler==0.10.0
decorator==4.4.2
dj-database-url==0.5.0
Django==3.0.4
django-background-tasks==1.2.5
django-celery-beat==2.0.0
django-ckeditor==5.9.0
django-comments-xtd==2.6.2
django-compat==1.0.15
django-contrib-comments==1.9.2
django-crispy-forms==1.9.2
django-cronjobs==0.2.3
django-heroku==0.3.1
django-js-asset==1.2.2
django-picklefield==3.0.1
django-q==1.3.2
django-redis==4.11.0
django-redis-cache==2.1.0
django-timezone-field==4.0
djangorestframework==3.11.0
docutils==0.16
Flask==1.1.1
google-api-core==1.20.0
google-api-python-client==1.9.3
google-auth==1.17.2
google-auth-httplib2==0.0.3
googleapis-common-protos==1.52.0
gunicorn==19.9.0
heroku==0.1.4
html5lib==1.0.1
httplib2==0.18.1
idna==2.8
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.1
joblib==0.15.1
kiwisolver==1.1.0
kombu==4.6.8
librosa==0.7.2
llvmlite==0.33.0
lxml==4.5.0
MarkupSafe==1.1.1
matplotlib==3.2.1
nltk==3.4.5
node==0.9.24
npm==0.1.1
numba==0.50.1
numpy==1.18.2
oauthlib==3.1.0
odict==1.7.0
optional-django==0.1.0
pandas==1.0.3
Pillow==7.2.0
plumber==1.5
protobuf==3.12.2
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
Pygments==2.6.1
PyJWT==1.7.1
pyOpenSSL==19.0.0
pyparsing==2.4.6
pypi==2.1
PySocks==1.7.0
python-crontab==2.5.1
python-dateutil==2.8.1
pytube==9.6.0
pytube3==9.6.4
pytz==2019.2
redis==3.4.1
requests==2.22.0
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.6
schedule==0.6.0
scikit-learn==0.23.1
scipy==1.5.0
selenium==3.141.0
six==1.12.0
SoundFile==0.10.3.post1
soupsieve==1.9.2
sqlparse==0.3.0
textblob==0.15.3
threadpoolctl==2.1.0
tweepy==3.8.0
twilio==6.42.0
typing-extensions==3.7.4.2
uritemplate==3.0.1
urllib3==1.24.2
vine==1.3.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.0
whitenoise==4.1.3
zipp==3.1.0
zope.component==4.6.1
zope.deferredimport==4.3.1
zope.deprecation==4.4.0
zope.event==4.4
zope.hookable==5.0.1
zope.interface==5.0.1
zope.lifecycleevent==4.3
zope.proxy==4.3.5
